// server/features/draft/types/initializerInferfaces.ts
export interface FearlessInitializerProps {
  fearlessCode: string;
  team1Code: string;
  team2Code: string;
  team1Name: string;
  team2Name: string;
  draftCount: number;
  initialTournamentCode?: string;
}

export interface FearlessFinishedProps {
  fearlessCode: string;
  fearlessComplete: boolean;
  team1Code: string;
  team2Code: string;
  team1Name: string;
  team2Name: string;
  draftCount: number;
}

// Holds the server's version of fearless state
// Includes the codes for each side
export interface FearlessStateServerProps {
  fearlessCode: string;
  fearlessComplete: boolean;
  team1Code: string;
  team2Code: string;
  team1Name: string;
  team2Name: string;
  draftCount: number;
  completedDrafts: number;
  currentDraft: string | null;
  currentBlueSide: string | null;
  currentRedSide: string | null;
  allPicks: string[];
  allBans: string[];
  team1Picks: string[];
  team2Picks: string[];
  team1Bans: string[];
  team2Bans: string[];
  draftLobbyCodes: string[];
  initialTournamentCode?: string;
}

export interface FearlessStateClientProps {
  fearlessCode: string;
  fearlessComplete: boolean;
  team1Name: string;
  team2Name: string;
  draftCount: number;
  completedDrafts: number;
  currentDraft: string | null;
  currentBlueSide: string | null;
  currentRedSide: string | null;
  allPicks: string[];
  allBans: string[];
  team1Picks: string[];
  team2Picks: string[];
  team1Bans: string[];
  team2Bans: string[];
  draftLobbyCodes: string[];
}
export interface DraftInitializerProps {
  lobbyCode: string;
  blueUser: string;
  redUser: string;
  blueDisplayName: string;
  redDisplayName: string;
  tournamentID: string | null;
  fearlessCode?: string;
}

export interface FearlessDraftInitializerProps {
  lobbyCode: string;
  blueUser: string;
  redUser: string;
  blueDisplayName: string;
  redDisplayName: string;
  tournamentID: string | null;
  fearlessCode: string;
}